Fix the proxy auth problem
